mcp_servers/gmail/requirements.txt
# Gmail MCP Server Dependencies
mcp>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# HTTP Server Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
